Refactor: move playlist components to /pages/playlist. Playlist header now a separate component.
