+++ src/app/App.tsx
@@ -7,7 +7,7 @@ import Content from '../components/Content'
 
 import Home from './pages/Home'
 import Search from './pages/Search'
-import Playlist from './pages/Playlist'
+import Playlist from './pages/playlist/Playlist'
 import Album from './pages/album/Album'
 import Artist from './pages/artist/Artist'
 import Profile from './pages/Profile'
